Replaced log4j with java logging, removed OWLManager references
In order to fit this as a module for OWLAPI, references to OWLManager
need to exist only in the test packages.

OWLAPI uses java logging, so replacing log4j to avoid client code 
having to set up both logging systems.
This might be reverted later, as it's not clear yet whether OWLAPI
should stick with java logging or move to a different API.